requirements.txt
Flask
validators
requests
beautifulsoup4
selenium
chardet
deep-translator
googlesearch-python
ollama
